requirements.txt
django>=1.1,<2
djangorestframework>=3.6,<4
django-cors-headers>=2.1,<3
django-allauth>=0.3,<1